refactor: 신고 후처리 이벤트 리스너 패키지 변경
