+++ space-d/src/main/java/com/dnd/spaced/core/admin/application/event/listener/exception/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.admin.application.event.exception;
+package com.dnd.spaced.core.admin.application.event.listener.exception;
 
 import com.dnd.spaced.global.exception.base.ReportServerException;
 import com.dnd.spaced.global.exception.code.CommentErrorCode;
